requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.0
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
redis==5.0.8
service-identity==24.1.0
setuptools==70.1.0
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.0.1
